Move remote config to its own module
Issues:
Fixes #<issueid>

Problem:

Analysis:

Tests:
